requirements.txt
django==3.2
requests==2.28.1
gunicorn==20.1.0
uvicorn==0.18.2
djangorestframework==3.13.1
phonenumbers==8.12.52
python-dotenv==0.20.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.2.2
psycopg2-binary==2.9.1